requirements.txt
python-telegram-bot==20.8
aiohttp==3.9.5
openpyxl==3.1.2
nest_asyncio